refactor: extract telemetry, replay, cloud, and edge modules to turn-core
Phase 17: 7 modules → astra-turn-core (+75 tests)

Modules extracted:
- turn/activity.rs (62 LOC) — session activity update planning
- turn/chat_turn_edge_profile.rs (128 LOC) — edge profile builder
- turn/sse_data_lines.rs (254 LOC) — SSE data line parser
- turn/agentic_turn_telemetry.rs (178 LOC) — turn telemetry helpers
- turn/session_replay_harness.rs (668 LOC) — session replay metrics
- cloud/compact_prompt.rs (545 LOC) — compaction prompt builder
- cloud/summary.rs (438 LOC) — summary generation with PTL retry

Changes:
- Added async-trait, reqwest deps to turn-core
- Moved MockSummaryClient from #[cfg(test)] to test_support module
  for cross-crate testing
- Fixed runtime compaction.rs to use test_support path
- Fixed clippy warnings across workspace (collapsible_if, field_reassign,
  derive Default, missing meta field in rmcp CreateElicitationResult)

Turn-core: 1558 tests. Runtime: 3244 + 5 pre-existing failures.
