+++ rust/crates/astra-skills/src/activation.rs
@@ -142,7 +142,7 @@ pub fn detect_triggers(skills: &[SkillManifest], message: &str) -> Vec<String> {
         }
     }
 
-    matches.sort_by(|a, b| b.1.cmp(&a.1));
+    matches.sort_by_key(|b| std::cmp::Reverse(b.1));
     matches.into_iter().map(|(name, _)| name).collect()
 }
