+++ rust/crates/astra-tools/src/tool_search.rs
@@ -132,7 +132,7 @@ pub fn tool_search(schemas: &[Value], args: &Value) -> String {
         })
         .collect();
 
-    scored.sort_by(|a, b| b.0.cmp(&a.0));
+    scored.sort_by_key(|b| std::cmp::Reverse(b.0));
 
     let matches: Vec<Value> = scored
         .into_iter()
