+++ rust/crates/astra-turn-core/Cargo.toml
@@ -10,12 +10,14 @@ astra-sandbox.workspace = true
 astra-services.workspace = true
 astra-text-utils.workspace = true
 astra-thin-client.workspace = true
+async-trait = { workspace = true }
 base64 = { workspace = true }
 chrono = { workspace = true }
 dashmap = "6.1.0"
 dirs = "6.0.0"
 fastrand = "2"
 regex = { workspace = true }
+reqwest = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 serde_yaml = "0.9.34"
