+++ rust/crates/runtime/src/plan/mod.rs
@@ -1,3 +1,3 @@
-pub use astra_plan::plan::*;
 pub use astra_plan::metrics;
 pub use astra_plan::performance;
+pub use astra_plan::plan::*;
